# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api SDE for mobile sde",
        "description": "Implementation of Swagger with in Laravel",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "https://haki44.eu/api",
            "description": "API SDE"
        }
    ],
    "paths": {
...
        "/boat-booking/{id}": {
            "get": {
                "tags": [
                    "Boat Booking"
                ],
                "summary": "Show one boat booking",
                "description": "Show one boat booking",
                "operationId": "ShowBoatBooking",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of boat booking",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Boat Booking"
                ],
                "summary": "Update one boat booking",
                "description": "Update one boat booking",
                "operationId": "UpdateBoatBooking",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of boat booking",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/boat": {
            "get": {
                "tags": [
                    "Boats"
                ],
                "summary": "Get List Of boats",
                "description": "Returns all boats",
                "operationId": "indexBoat",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Boats"
                ],
                "summary": "Store boat",
                "description": "Store boat",
                "operationId": "storeBoat",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/boat/{id}": {
            "get": {
                "tags": [
                    "Boats"
                ],
                "summary": "Get details Of boat",
                "description": "Returns details Of boat",
                "operationId": "showBoat",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of boat",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Boats"
                ],
                "summary": "Update one boat",
                "description": "Update one boat",
                "operationId": "updateBoat",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of boat",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Boats"
                ],
                "summary": "Delete one boat",
                "description": "Delete one boat",
                "operationId": "deleteBoat",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of boat",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    }
}